moving readMarkdownFolder into it's own file
